Rewriting to avoid inconsistency of coordinates
In some parts of the text and figures the "actual" coordinates (ia, jb) were used, whereas in others the normalized coordinates (i,j) were used. They were even mixed in the same sentence of figure. This has been unified in the text for consistency. The actual coordinates are now used. The choice had to be this, otherwise the intuition of Euclidean distance was lost. Pending: correct labels in Figures 6 and 8 according to this